cals/cals-sdk/requirements.txt
cvat-sdk==2.10.0
cvat-sdk[pytorch]
easydict==1.10
numpy==1.26.2
Pillow==10.1.0
python-dotenv==1.0.0
# pytorch-lightning
requests==2.28.2
torch==2.0.1
torchvision==0.15.2
# tqdm